refactor: split routes into controllers, validators, and pure wiring
Route files now contain ONLY middleware and route registration:
  app.get('/api/org', getOrg(deps))

Handler logic extracted to src/presentation/controllers/ (16 files).
Zod schemas extracted to src/presentation/validators/ (9 files).

Structure:
  src/presentation/
  ├── routes/        Pure wiring (middleware + route registration)
  ├── controllers/   Handler functions (parse, call use case, respond)
  └── validators/    Zod schemas

Route files reduced from 30-180 lines to 15-35 lines each.
Zero functional changes. 439/439 tests pass.
